backend/requirements.txt
fastapi
uvicorn
python-dotenv
azure-ai-textanalytics
azure-cognitiveservices-speech
requests>=2.28.0
